Update contract dependencies and test setup logic for platform-contracts 0.2.0 and token-contracts 2.0.0
